squash: further simplify generateSubsteps by moving getter factories out
