+++ protocol-designer/src/top-selectors/substeps.js
@@ -8,12 +8,28 @@ import {selectors as fileDataSelectors} from '../file-data'
 import {allWellContentsForSteps} from './well-contents'
 
 import {
-  generateSubsteps
+  generateSubsteps,
+  type GetIngreds
 } from '../steplist/generateSubsteps' // TODO Ian 2018-04-11 move generateSubsteps closer to this substeps.js file?
 
 import type {Selector} from '../types'
 import type {StepIdType} from '../form-types'
 import type {SubstepItemData} from '../steplist/types'
+import type {WellContentsByLabware} from './well-contents'
+
+const getIngredsFactory = (
+  wellContentsByLabware: WellContentsByLabware,
+  ingredNames: {[ingredId: string]: string}
+): GetIngreds => (labware, well) => {
+  const wellContents = (wellContentsByLabware &&
+    wellContentsByLabware[labware] &&
+    wellContentsByLabware[labware][well])
+
+  return wellContents.groupIds.map(id => ({
+    id: id,
+    name: ingredNames[id]
+  })) || []
+}
 
 type AllSubsteps = {[StepIdType]: ?SubstepItemData}
 export const allSubsteps: Selector<AllSubsteps> = createSelector(
@@ -42,9 +58,8 @@ export const allSubsteps: Selector<AllSubsteps> = createSelector(
       const substeps = generateSubsteps(
         validatedForms[stepId],
         allPipetteData,
-        allLabwareTypes,
-        ingredNames,
-        _allWellContentsForSteps[timelineIndex],
+        (labwareId: string) => allLabwareTypes[labwareId],
+        getIngredsFactory(_allWellContentsForSteps[timelineIndex], ingredNames),
         robotState,
         stepId
       )
